+++ pyobjc-framework-IOBluetoothUI/PyObjCTest/test_iobluetoothobjectpushuicontroller.py
@@ -1,0 +1,10 @@
+from PyObjCTools.TestSupport import TestCase
+
+import IOBluetoothUI
+
+
+class TestIOBluetoothObjectPushUIController(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(
+            IOBluetoothUI.IOBluetoothObjectPushUIController.isTransferInProgress
+        )