+++ pyobjc-framework-IOBluetoothUI/PyObjCTest/test_iobluetoothpairingcontroller.py
@@ -1,0 +1,10 @@
+from PyObjCTools.TestSupport import TestCase
+
+import IOBluetoothUI
+
+
+class TestIOBluetoothPairingController(TestCase):
+    def test_methods(self):
+        self.assertArgIsIn(
+            IOBluetoothUI.IOBluetoothPairingController.setSearchAttributes_, 0
+        )