+++ docs/changelog.rst
@@ -336,7 +336,7 @@ improvements as described below).
   ``objc.NSDecimal(0)`` was interpreted as true-ish in boolean contexts, it now
   is interpreted as false-ish just like other number types.
 
-* Add bindings for the IOBluetooth framework
+* :issue:`381`: Add bindings for the IOBluetooth and IOBluetoothUI frameworks
 
 Version 9.0.1
 -------------